refactor: PR tweaks
docs: Regenerate after mdoc upgrade + fix in jvm.md
