+++ docs/flyway.md
@@ -4,7 +4,7 @@
 
 `libraryDependencies += "com.avast" %% "sst-flyway" % "<VERSION>"`
 
-This module initializes `Flyway` which can be used to do automated SQL DB migrations. See the [documentation of Flyway](https://flywaydb.org/documentation/)
+This module initializes `Flyway` which can be used to do automated SQL DB migrations. See the [documentation of Flyway](https://flywaydb.org/documentation/) 
 on how to go about that.
 
 The method `make` requires `javax.sql.DataSource` which you can for example obtain from `doobie-hikari` module:
@@ -22,4 +22,3 @@ for {
   _ <- Resource.liftF(Task.effect(flyway.migrate()))
 } yield ()
 ```
-
