+++ docs/monix-catnap.md
@@ -12,7 +12,7 @@ You can use `CircuitBreakerModule` to instantiate and configure a circuit breake
 monitoring of the circuit breaker. There is an implementation for Micrometer in `sst-monix-catnap-micrometer` module.
 
 All of this is tied with http4s HTTP client in the `sst-http4s-client-monix-catnap` module so in practice you want to use
-`Http4sClientCircuitBreakerModule` which wraps any `Client[F]` with a `CircuitBreaker` and adds logging/metrics. However
+`Http4sClientCircuitBreakerModule` which wraps any `Client[F]` with a `CircuitBreaker` and adds logging/metrics. However 
 the most important feature of the enriched circuit breaker is that any HTTP failure (response `5xx`) is converted to an exception internally
 which triggers the circuit breaking mechanism. Failing server is not overloaded by more requests and we do not have to wait for the response
 if the server is failing anyway.
