+++ monix-catnap-micrometer/src/main/scala/com/avast/sst/monix/catnap/micrometer/MicrometerCircuitBreakerMetricsModule.scala
@@ -1,15 +1,44 @@
 package com.avast.sst.monix.catnap.micrometer
 
+import java.util.concurrent.atomic.AtomicInteger
+
 import cats.effect.Sync
 import com.avast.sst.monix.catnap.CircuitBreakerMetrics
-import com.avast.sst.monix.catnap.micrometer.MicrometerCircuitBreakerMetrics.Impl
+import com.avast.sst.monix.catnap.CircuitBreakerMetrics.State
+import com.avast.sst.monix.catnap.CircuitBreakerMetrics.State.{Closed, HalfOpen, Open}
 import io.micrometer.core.instrument.MeterRegistry
 
 object MicrometerCircuitBreakerMetricsModule {
 
   /** Makes [[com.avast.sst.monix.catnap.CircuitBreakerMetrics]] from [[io.micrometer.core.instrument.MeterRegistry]]. */
   def make[F[_]: Sync](name: String, meterRegistry: MeterRegistry): F[CircuitBreakerMetrics[F]] = {
-    Sync[F].delay(new Impl(name, meterRegistry))
+    Sync[F].delay(new MicrometerCircuitBreakerMetrics(name, meterRegistry))
+  }
+
+  private class MicrometerCircuitBreakerMetrics[F[_]: Sync](name: String, meterRegistry: MeterRegistry) extends CircuitBreakerMetrics[F] {
+
+    private val F = Sync[F]
+
+    private[this] val CircuitOpened = -1
+    private[this] val CircuitHalfOpened = 0
+    private[this] val CircuitClosed = 1
+
+    private val accepted = meterRegistry.counter(s"circuit-breaker.$name.accepted")
+    private val rejected = meterRegistry.counter(s"circuit-breaker.$name.rejected")
+    private val circuitState = meterRegistry.gauge[AtomicInteger](s"circuit-breaker.$name.state", new AtomicInteger(CircuitClosed))
+
+    override def increaseAccepted: F[Unit] = F.delay(accepted.increment())
+
+    override def increaseRejected: F[Unit] = F.delay(rejected.increment())
+
+    override def setState(state: State): F[Unit] = {
+      state match {
+        case Closed   => F.delay(circuitState.set(CircuitClosed))
+        case Open     => F.delay(circuitState.set(CircuitOpened))
+        case HalfOpen => F.delay(circuitState.set(CircuitHalfOpened))
+      }
+    }
+
   }
 
 }
